3次元における2点間の中点

+++ Assets/Script/Math_and_Physics/Math_and_PhysicsLib.cs
@@ -24,7 +24,7 @@ public class Math_and_PhysicsLib
     public static extern IntPtr find2DMidPoint(float[] P1, float[] P2);
 
     [DllImport("NativePluginSample")]
-    public static extern float find3DMidPoint(float[] P1, float[] P2);
+    public static extern IntPtr find3DMidPoint(float[] P1, float[] P2);
 
     public struct sphere { };
